LayoutTests/media/media-vp8-webm-with-preload-expected.txt


EXPECTED (video.readyState <= '1') OK
EXPECTED (video.readyState == '1') OK
RUN(video.play())
EVENT(playing)
RUN(video.currentTime = video.duration - 0.02)
